+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoRuntime.java
@@ -29,6 +29,7 @@ import android.util.Log;
 import androidx.annotation.AnyThread;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.OptIn;
 import androidx.annotation.StringDef;
 import androidx.annotation.UiThread;
 import androidx.lifecycle.Lifecycle;
@@ -247,7 +248,10 @@ public final class GeckoRuntime implements Parcelable {
   private GeckoRuntimeSettings mSettings;
   private Delegate mDelegate;
   private ServiceWorkerDelegate mServiceWorkerDelegate;
+
+  @OptIn(markerClass = ExperimentalGeckoViewApi.class)
   private GeckoPreferenceController.Observer.Delegate mPreferencesObserverDelegate;
+
   private WebNotificationDelegate mNotificationDelegate;
   private ActivityDelegate mActivityDelegate;
   private OrientationController mOrientationController;
@@ -315,6 +319,7 @@ public final class GeckoRuntime implements Parcelable {
     }
   }
 
+  @OptIn(markerClass = ExperimentalGeckoViewApi.class)
   private final BundleEventListener mEventListener =
       new BundleEventListener() {
         @Override
@@ -777,6 +782,7 @@ public final class GeckoRuntime implements Parcelable {
    *
    * @param delegate The delegate to set on the runtime.
    */
+  @ExperimentalGeckoViewApi
   @AnyThread
   public void setPreferencesObserverDelegate(
       @Nullable final GeckoPreferenceController.Observer.Delegate delegate) {
@@ -789,6 +795,7 @@ public final class GeckoRuntime implements Parcelable {
    *
    * @return The {@link GeckoPreferenceController.Observer.Delegate} set on this runtime.
    */
+  @ExperimentalGeckoViewApi
   @AnyThread
   public @Nullable GeckoPreferenceController.Observer.Delegate getPreferencesObserverDelegate() {
     return mPreferencesObserverDelegate;